docs : move StyledFooterHeading to Footer/UI folder
